requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo
scikit-learn
dill
pyyaml
xgboost
fastapi
uvicorn
ipykernel
apache-airflow
mlflow
dvc
jinja2
python-multipart